Refactor window layer implementation.
Introduced the concept of a window type (level/layer is overloaded) and
deprecated the notion of 'interaction enabled'. It is now controlled by
the window type.

The window type is now appended to the URL, which makes it
distinguishable in the debug console and opens the door for selective
loading of widgets in the future.