Randomize local server port
Make the port of the local server harder to guess, which will help prevent
naïve exploitation of the privilege escalation threat posed by the `/run/`
endpoint.  This can be circumvented by inspecting the command line of the
process, or scanning all the local ports, to discover the correct port;
however, it is better than nothing.