bugfix: Only call shutdown once at a time
Sometimes `shutdown:true` is called when modifying settings through the
preferences window.  However, `shutdown` (effectively `shutdown:false`) is
called when the node process exits, which results in `keepServerAlive` being
set to `NO`.  Since `keepServerAlive` is always `NO` at the time of the check,
the node process is never restarted, leaving the program in a crippled state.

Create a flag to indicate whether the shutdown flow is ongoing.  If the flag is
true, then `shutdown:keepAlive` can exit immediately without any adverse side
effects.